Remove `seq` from ThreeDeeRender and ThreeDimensionalViz publish messages
When publishing messages to a ROS 2 source (rosbridge), the `seq` field would cause the rosbridge to fail because headers in ROS2 do not have a `seq` field.

This change removes the `seq` field from ThreeDeeRender and ThreeDimensionalViz so those messages are valid to publish.

Fixes: #2834
